cli demos: convert demo_pipeline_voltage_imaging to "app"

+++ demos/general/demo_pipeline_voltage_imaging.py
@@ -7,11 +7,11 @@ objects and call the relevant functions. See inside for detail.
 Dataset courtesy of Karel Svoboda Lab (Janelia Research Campus).
 author: @caichangjia
 """
-#%%
+
+import argparse
 import cv2
 import glob
 import h5py
-from IPython import get_ipython
 import logging
 import matplotlib.pyplot as plt
 import numpy as np
@@ -22,45 +22,41 @@ try:
 except:
     pass
 
-try:
-    if __IPYTHON__:
-        ipython = get_ipython()
-        ipython.run_line_magic('load_ext', 'autoreload')
-        ipython.run_line_magic('autoreload', '2')
-        ipython.run_line_magic('matplotlib', 'qt')
-except NameError:
-    pass
-
 import caiman as cm
 from caiman.motion_correction import MotionCorrect
 from caiman.paths import caiman_datadir
 from caiman.source_extraction.volpy import utils
 from caiman.source_extraction.volpy.volparams import volparams
 from caiman.source_extraction.volpy.volpy import VOLPY
-from caiman.summary_images import local_correlations_movie_offline
-from caiman.summary_images import mean_image
+from caiman.summary_images import local_correlations_movie_offline, mean_image
 from caiman.utils.utils import download_demo, download_model
 
-# %%
-# Set up the logger (optional); change this if you like.
-# You can log to a file using the filename parameter, or make the output more
-# or less verbose by setting level to logging.DEBUG, logging.INFO,
-# logging.WARNING, or logging.ERROR
-logging.basicConfig(format=
-                    "%(relativeCreated)12d [%(filename)s:%(funcName)20s():%(lineno)s]" \
-                    "[%(process)d] %(message)s",
-                    level=logging.INFO)
 
-# %%
 def main():
-    pass  # For compatibility between running using an IDE and the CLI
+    cfg = handle_args()
 
-    # %%  Load demo movie and ROIs
-    fnames = download_demo('demo_voltage_imaging.hdf5', 'volpy')  # file path to movie file (will download if not present)
-    path_ROIs = download_demo('demo_voltage_imaging_ROIs.hdf5', 'volpy')  # file path to ROIs file (will download if not present)
+    if cfg.logfile:
+        logging.basicConfig(format=
+            "%(relativeCreated)12d [%(filename)s:%(funcName)20s():%(lineno)s][%(process)d] %(message)s",
+            level=logging.WARNING,
+            filename=cfg.logfile)
+        # You can make the output more or less verbose by setting level to logging.DEBUG, logging.INFO, logging.WARNING, or logging.ERROR
+    else:
+        logging.basicConfig(format=
+            "%(relativeCreated)12d [%(filename)s:%(funcName)20s():%(lineno)s][%(process)d] %(message)s",
+            level=logging.WARNING)
+
+    if cfg.input is None:
+        # If no input is specified, use sample data, downloading if necessary
+        fnames    = [download_demo('demo_voltage_imaging.hdf5', 'volpy')] # XXX do we need to use a separate directory?
+        path_ROIs = [download_demo('demo_voltage_imaging_ROIs.hdf5', 'volpy')]
+    else:
+        fnames = cfg.input
+        path_ROIs = cfg.pathinput
+
+    #  Load demo movie and ROIs
     file_dir = os.path.split(fnames)[0]
     
-    #%% dataset dependent parameters
     # dataset dependent parameters
     fr = 400                                        # sample rate of the movie
 
@@ -74,33 +70,29 @@ def main():
     max_deviation_rigid = 3                         # maximum deviation allowed for patch with respect to rigid shifts
     border_nan = 'copy'
 
-    opts_dict = {
-        'fnames': fnames,
-        'fr': fr,
-        'pw_rigid': pw_rigid,
-        'max_shifts': max_shifts,
-        'gSig_filt': gSig_filt,
-        'strides': strides,
-        'overlaps': overlaps,
-        'max_deviation_rigid': max_deviation_rigid,
-        'border_nan': border_nan
-    }
+    params_dict = {'fnames': fnames,
+                   'fr': fr,
+                   'pw_rigid': pw_rigid,
+                   'max_shifts': max_shifts,
+                   'gSig_filt': gSig_filt,
+                   'strides': strides,
+                   'overlaps': overlaps,
+                   'max_deviation_rigid': max_deviation_rigid,
+                   'border_nan': border_nan}
 
-    opts = volparams(params_dict=opts_dict)
+    opts = volparams(params_dict=params_dict)
 
-    # %% play the movie (optional)
+    # play the movie (optional)
     # playing the movie using opencv. It requires loading the movie in memory.
     # To close the movie press q
-    display_images = False
-    if display_images:
+    if not cfg.no_play:
         m_orig = cm.load(fnames)
         ds_ratio = 0.2
         moviehandle = m_orig.resize(1, 1, ds_ratio)
         moviehandle.play(q_max=99.5, fr=40, magnification=4)
 
-    # %% start a cluster for parallel processing
-    c, dview, n_processes = cm.cluster.setup_cluster(
-        backend='multiprocessing', n_processes=None, single_thread=False)
+    # start a cluster for parallel processing
+    c, dview, n_processes = cm.cluster.setup_cluster(backend=cfg.cluster_backend)
 
     # %%% MOTION CORRECTION
     # first we create a motion correction object with the specified parameters
@@ -115,8 +107,8 @@ def main():
         mc.mmap_file = [os.path.join(file_dir, mc_list[0])]
         print(f'reuse previously saved motion corrected file:{mc.mmap_file}')
 
-    # %% compare with original movie
-    if display_images:
+    # compare with original movie
+    if not cfg.no_play:
         m_orig = cm.load(fnames)
         m_rig = cm.load(mc.mmap_file)
         ds_ratio = 0.2
@@ -124,7 +116,7 @@ def main():
                                       m_rig.resize(1, 1, ds_ratio)], axis=2)
         moviehandle.play(fr=40, q_max=99.5, magnification=4)  # press q to exit
 
-    # %% MEMORY MAPPING
+    # MEMORY MAPPING
     do_memory_mapping = True
     if do_memory_mapping:
         border_to_0 = 0 if mc.border_nan == 'copy' else mc.border_to_0
@@ -139,9 +131,9 @@ def main():
         mmap_list = [file for file in os.listdir(file_dir) if 
                      ('memmap_' + os.path.splitext(os.path.split(fnames)[-1])[0]) in file]
         fname_new = os.path.join(file_dir, mmap_list[0])
-        print(f'reuse previously saved memory mapping file:{fname_new}')
+        print(f'reuse previously saved memory mapping file: {fname_new}')
     
-    # %% SEGMENTATION
+    # SEGMENTATION
     # create summary images
     img = mean_image(mc.mmap_file[0], window = 1000, dview=dview)
     img = (img-np.mean(img))/np.std(img)
@@ -159,7 +151,7 @@ def main():
     axs[0].imshow(summary_images[0]); axs[1].imshow(summary_images[2])
     axs[0].set_title('mean image'); axs[1].set_title('corr image');
 
-    #%% methods for segmentation
+    # methods for segmentation
     methods_list = ['manual_annotation',       # manual annotations need prepared annotated datasets in the same format as demo_voltage_imaging_ROIs.hdf5 
                     'maskrcnn',                # Mask R-CNN is a convolutional neural network trained for detecting neurons in summary images
                     'gui_annotation']          # use VolPy GUI to correct outputs of Mask R-CNN or annotate new datasets 
@@ -177,7 +169,6 @@ def main():
 
     elif method == 'gui_annotation':
         # run volpy_gui.py file in the caiman/source_extraction/volpy folder
-        # or run the following in the ipython:  %run volpy_gui.py
         gui_ROIs =  caiman_datadir() + '/example_movies/volpy/gui_roi.hdf5'
         with h5py.File(gui_ROIs, 'r') as fl:
             ROIs = fl['mov'][()]
@@ -186,12 +177,12 @@ def main():
     axs[0].imshow(summary_images[0]); axs[1].imshow(ROIs.sum(0))
     axs[0].set_title('mean image'); axs[1].set_title('masks')
         
-    # %% restart cluster to clean up memory
+    # restart cluster to clean up memory
     cm.stop_server(dview=dview)
-    c, dview, n_processes = cm.cluster.setup_cluster(
-        backend='multiprocessing', n_processes=None, single_thread=False, maxtasksperchild=1)
+    c, dview, n_processes = cm.cluster.setup_cluster(backend=cfg.cluster_backend)
 
-    # %% parameters for trace denoising and spike extraction
+
+    # parameters for trace denoising and spike extraction
     ROIs = ROIs                                   # region of interests
     index = list(range(len(ROIs)))                # index of neurons
     weights = None                                # if None, use ROIs for initialization; to reuse weights check reuse weights block 
@@ -232,27 +223,26 @@ def main():
                'weight_update': weight_update,
                'n_iter': n_iter}
 
-    opts.change_params(params_dict=opts_dict);          
+    opts.change_params(params_dict=params_dict);          
 
-    #%% TRACE DENOISING AND SPIKE DETECTION
+    # TRACE DENOISING AND SPIKE DETECTION
     vpy = VOLPY(n_processes=n_processes, dview=dview, params=opts)
     vpy.fit(n_processes=n_processes, dview=dview);
    
-    #%% visualization
-    display_images = True
-    if display_images:
+    # visualization
+    if not cfg.no_play:
         print(np.where(vpy.estimates['locality'])[0])    # neurons that pass locality test
         idx = np.where(vpy.estimates['locality'] > 0)[0]
         utils.view_components(vpy.estimates, img_corr, idx)
     
-    #%% reconstructed movie 
+    # reconstructed movie 
     # note the negative spatial weights are cutoff    
-    if display_images:
+    if not cfg.no_play:
         mv_all = utils.reconstructed_movie(vpy.estimates.copy(), fnames=mc.mmap_file,
                                            idx=idx, scope=(0,1000), flip_signal=flip_signal)
         mv_all.play(fr=40, magnification=3)
     
-    #%% save the result in .npy format 
+    # save the result in .npy format 
     save_result = True
     if save_result:
         vpy.estimates['ROIs'] = ROIs
@@ -260,7 +250,7 @@ def main():
         save_name = f'volpy_{os.path.split(fnames)[1][:-5]}_{threshold_method}'
         np.save(os.path.join(file_dir, save_name), vpy.estimates)
         
-    #%% reuse weights 
+    # reuse weights 
     # set weights = reuse_weights in opts_dict dictionary
     estimates = np.load(os.path.join(file_dir, save_name+'.npy'), allow_pickle=True).item()
     reuse_weights = []
@@ -270,15 +260,24 @@ def main():
         plt.figure(); plt.imshow(w);plt.colorbar(); plt.show()
         reuse_weights.append(w)
     
-    # %% STOP CLUSTER and clean up log files
+    # Stop the cluster and clean up log files
     cm.stop_server(dview=dview)
-    log_files = glob.glob('*_LOG_*')
-    for log_file in log_files:
-        os.remove(log_file)
 
-# %%
-# This is to mask the differences between running this demo in an IDE
-# versus from the CLI
-if __name__ == "__main__":
-    main()
+    if not cfg.keep_logs:
+        log_files = glob.glob('*_LOG_*')
+        for log_file in log_files:
+            os.remove(log_file)
+
+def handle_args():
+    parser = argparse.ArgumentParser(description="Demonstrate voltage imaging pipeline")
+    parser.add_argument("--keep_logs",  action="store_true", help="Keep temporary logfiles")
+    parser.add_argument("--no_play",    action="store_true", help="Do not display results")
+    parser.add_argument("--cluster_backend", default="multiprocessing", help="Specify multiprocessing, ipyparallel, or single to pick an engine")
+    parser.add_argument("--input", action="append", help="File(s) to work on, provide multiple times for more files")
+    parser.add_argument("--pathinput", action="append", help="Path input file(s) to work on, provide multiple times for more files")
+    parser.add_argument("--logfile",    help="If specified, log to the named file")
+    return parser.parse_args()
+
+########
+main()
